fix: fix incorrect skeleton display
